+++ src/test/java/com/uber/h3core/TestH3CoreLoader.java
@@ -49,17 +49,17 @@ public class TestH3CoreLoader {
 
     @Test
     public void testDetectArch() {
-        assertEquals("x64", H3CoreLoader.detectArch("amd64"));
-        assertEquals("x64", H3CoreLoader.detectArch("x86_64"));
-        assertEquals("x64", H3CoreLoader.detectArch("x64"));
+        assertEquals(H3CoreLoader.ARCH_X64, H3CoreLoader.detectArch("amd64"));
+        assertEquals(H3CoreLoader.ARCH_X64, H3CoreLoader.detectArch("x86_64"));
+        assertEquals(H3CoreLoader.ARCH_X64, H3CoreLoader.detectArch("x64"));
 
-        assertEquals("x86", H3CoreLoader.detectArch("x86"));
-        assertEquals("x86", H3CoreLoader.detectArch("i386"));
-        assertEquals("x86", H3CoreLoader.detectArch("i486"));
-        assertEquals("x86", H3CoreLoader.detectArch("i586"));
-        assertEquals("x86", H3CoreLoader.detectArch("i686"));
-        assertEquals("x86", H3CoreLoader.detectArch("i786"));
-        assertEquals("x86", H3CoreLoader.detectArch("i886"));
+        assertEquals(H3CoreLoader.ARCH_X86, H3CoreLoader.detectArch("x86"));
+        assertEquals(H3CoreLoader.ARCH_X86, H3CoreLoader.detectArch("i386"));
+        assertEquals(H3CoreLoader.ARCH_X86, H3CoreLoader.detectArch("i486"));
+        assertEquals(H3CoreLoader.ARCH_X86, H3CoreLoader.detectArch("i586"));
+        assertEquals(H3CoreLoader.ARCH_X86, H3CoreLoader.detectArch("i686"));
+        assertEquals(H3CoreLoader.ARCH_X86, H3CoreLoader.detectArch("i786"));
+        assertEquals(H3CoreLoader.ARCH_X86, H3CoreLoader.detectArch("i886"));
 
         assertEquals("anything", H3CoreLoader.detectArch("anything"));
         assertEquals("i986", H3CoreLoader.detectArch("i986"));