+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The public API of this library consists of the public functions declared in
 file [H3Core.java](./src/main/java/com/uber/h3core/H3Core.java), and support
 for the Linux x64 and Darwin x64 platforms.
 
-## Unreleased
+## [3.0.2] - 2018-05-08
 ### Fixed
 - Fixed memory management in polyfill with multiple holes.
 - Fixed build on Linux without Docker.